do helicone modification just before running the example file
